correct size for directed graph where we check if matrix is symetrical
